requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz==0.14.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
batchbald-redux==2.0.5
bcrypt==4.0.1
beautifulsoup4==4.11.1
blackhc.project==2.6.0
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
cloudpickle==2.2.0
conda==4.3.16
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distrax==0.1.2
dm-tree==0.1.7
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
firebase-admin==6.0.1
fonttools==4.38.0
fs==2.4.16
fs.sshfs==1.0.1
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
graphviz==0.20.1
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jax==0.3.25
jaxlib==0.3.25
jaxopt==0.5.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
laaos==2.1.1
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
netCDF4==1.6.2
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.4
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.3
prometheus-client==0.15.0
prompt-toolkit==3.0.32
property-cached==1.6.4
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PyNaCl==1.5.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.0
tensorflow-probability==0.18.0
terminado==0.17.0
TexSoup==0.3.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toma==1.1.0
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
umap-learn==0.5.3
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
xarray==2022.11.0
xarray-einstats==0.3.0
yarg==0.1.9
zipp==3.10.0
